requirements.txt
Django==3.1.2
python-decouple==3.3
pytz==2020.1
requests==2.24.0
gunicorn==20.0.4
